pyproject.toml
[project]
name = "kanitts-vllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.3.1",
    "accelerate==1.11.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "alembic==1.17.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.11.0",
    "astor==0.8.1",
    "asttokens==3.0.0",
    "async-timeout==5.0.1",
    "attrdict==2.0.1",
    "attrs==25.4.0",
    "audioread==3.0.1",
    "backports-datetime-fromisoformat==2.0.3",
    "bitsandbytes==0.46.0",
    "blake3==1.0.8",
    "braceexpand==0.1.7",
    "cachetools==6.2.1",
    "cbor2==5.7.0",
    "cdifflib==1.2.6",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorlog==6.10.1",
    "compressed-tensors==0.11.0",
    "contourpy==1.3.2",
    "cupy-cuda12x==13.6.0",
    "cycler==0.12.1",
    "cytoolz==1.1.0",
    "datasets==4.2.0",
    "decorator==5.2.1",
    "depyf==0.19.0",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docopt==0.6.2",
    "editdistance==0.8.1",
    "einops==0.8.1",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "executing==2.2.1",
    "fastapi==0.119.1",
    "fastapi-cli==0.0.14",
    "fastapi-cloud-cli==0.3.1",
    "fastrlock==0.8.3",
    "fiddle==0.3.0",
    "filelock==3.20.0",
    "flash-attn==2.8.3",
    "fonttools==4.60.1",
    "frozendict==2.4.6",
    "frozenlist==1.8.0",
    "fsspec==2024.12.0",
    "future==1.0.0",
    "gguf==0.17.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "graphviz==0.21",
    "greenlet==3.2.4",
    "grpcio==1.75.1",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.35.3",
    "hydra-core==1.3.2",
    "idna==3.11",
    "indic-numtowords==1.1.0",
    "inflect==7.5.0",
    "interegular==0.3.3",
    "intervaltree==3.1.0",
    "ipython==8.37.0",
    "janome==0.5.0",
    "jedi==0.19.2",
    "jieba==0.42.1",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "jiwer==3.1.0",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kaldi-python-io==1.2.2",
    "kiwisolver==1.4.9",
    "kornia==0.8.1",
    "kornia-rs==0.1.9",
    "lark==1.2.2",
    "lazy-loader==0.4",
    "levenshtein==0.27.1",
    "lhotse==1.31.1",
    "libcst==1.8.5",
    "librosa==0.11.0",
    "lightning==2.4.0",
    "lightning-utilities==0.15.2",
    "lilcom==1.8.1",
    "llguidance==0.7.30",
    "llvmlite==0.44.0",
    "lm-format-enforcer==0.11.3",
    "loguru==0.7.3",
    "mako==1.3.10",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==4.0.1",
    "matplotlib==3.10.7",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mediapy==1.1.6",
    "mistral-common==1.8.5",
    "ml-dtypes==0.5.3",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "msgspec==0.19.0",
    "multidict==6.7.0",
    "multiprocess==0.70.16",
    "nemo-text-processing==1.1.0",
    "nemo-toolkit==2.5.0",
    "networkx==3.4.2",
    "ninja==1.13.0",
    "nltk==3.9.2",
    "num2words==0.5.14",
    "numba==0.61.2",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.9.1.4",
    "nvidia-cuda-cupti-cu12==12.9.79",
    "nvidia-cuda-nvrtc-cu12==12.9.86",
    "nvidia-cuda-runtime-cu12==12.9.79",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.4.1.4",
    "nvidia-cufile-cu12==1.14.1.1",
    "nvidia-curand-cu12==10.3.10.19",
    "nvidia-cusolver-cu12==11.7.5.82",
    "nvidia-cusparse-cu12==12.5.10.65",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.9.86",
    "nvidia-nvtx-cu12==12.9.79",
    "omegaconf==2.3.0",
    "onnx==1.19.1",
    "openai==2.6.0",
    "openai-harmony==0.0.4",
    "opencv-python-headless==4.12.0.88",
    "optuna==4.5.0",
    "outlines-core==0.2.11",
    "packaging==24.2",
    "pandas==2.3.3",
    "parso==0.8.5",
    "partial-json-parser==0.2.1.1.post6",
    "peft==0.17.1",
    "pexpect==4.9.0",
    "pillow==12.0.0",
    "plac==1.4.5",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "prometheus-client==0.23.1",
    "prometheus-fastapi-instrumentator==7.1.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==5.29.5",
    "psutil==7.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "py-cpuinfo==9.0.0",
    "pyannote-core==5.0.0",
    "pyannote-database==5.1.3",
    "pyannote-metrics==3.2.1",
    "pyarrow==21.0.0",
    "pybase64==1.4.2",
    "pybind11==3.0.1",
    "pycountry==24.6.1",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-extra-types==2.10.6",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "pyloudnorm==0.1.1",
    "pynini==2.1.6.post1",
    "pyparsing==3.2.5",
    "pypinyin==0.55.0",
    "pypinyin-dict==0.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.20",
    "pytorch-lightning==2.5.5",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rapidfuzz==3.14.1",
    "ray==2.50.1",
    "referencing==0.37.0",
    "regex==2025.10.22",
    "requests==2.32.5",
    "resampy==0.4.3",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.1",
    "rpds-py==0.27.1",
    "ruamel-yaml==0.18.15",
    "ruamel-yaml-clib==0.2.14",
    "sacremoses==0.1.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "sentencepiece==0.2.1",
    "sentry-sdk==2.42.1",
    "setproctitle==1.3.7",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soundfile==0.13.1",
    "sox==1.5.0",
    "soxr==1.0.0",
    "sqlalchemy==2.0.44",
    "stack-data==0.6.3",
    "starlette==0.48.0",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "termcolor==3.1.0",
    "text-unidecode==1.3",
    "texterrors==0.5.1",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tomli==2.3.0",
    "toolz==1.1.0",
    "torch==2.8.0+cu129",
    "torchaudio==2.8.0+cu129",
    "torchmetrics==1.8.2",
    "torchvision==0.23.0+cu129",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.1",
    "triton==3.4.0",
    "typeguard==4.4.4",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "vllm==0.11.0",
    "wandb==0.22.2",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "webdataset==1.0.2",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "wget==3.2",
    "whisper-normalizer==0.1.12",
    "wrapt==2.0.0",
    "xformers==0.0.32.post1",
    "xgrammar==0.1.25",
    "xxhash==3.6.0",
    "yarl==1.22.0",
]

